Move findUnusedPort into module
