+++ lib/start-server.js
@@ -22,7 +22,7 @@ const chalk = require('chalk');
 const fs = require('fs');
 const program = require('commander');
 const openBrowser = require('react-dev-utils/openBrowser');
-const tcpPortUsed = require('tcp-port-used');
+const findUnusedPort = require('./findUnusedPort.js');
 
 const CWD = process.cwd();
 const env = require('./server/env.js');
@@ -46,30 +46,6 @@ program
   .option('--port <number>', 'Specify port number')
   .option('--no-watch', 'Toggle live reload file watching')
   .parse(process.argv);
-
-function findUnusedPort(initalPort) {
-  const maxAttempts = 10;
-  let numAttempts = 0;
-
-  const checkPort = port => {
-    const host = 'localhost';
-    return tcpPortUsed.check(port, host).then(inUse => {
-      if (!inUse) {
-        return Promise.resolve(port);
-      }
-
-      numAttempts++;
-      if (numAttempts === maxAttempts) {
-        return Promise.reject(
-          new Error('cannot find unused port: max attempts exceeded')
-        );
-      }
-      return checkPort(port + 1);
-    });
-  };
-
-  return checkPort(initalPort);
-}
 
 function startLiveReloadServer() {
   const initalLiveReloadPort = 35729;
